Moved geometries into their specific folders for bifrost, loki, and miracles. Fixed tests.
