+++ src/efu/test/ServerTest.cpp
@@ -50,7 +50,7 @@ void client_thread(int command) {
 class TestDetector : public Detector {
 public:
   explicit TestDetector(BaseSettings settings)
-      : Detector("No name", settings){};
+      : Detector(settings){};
   ~TestDetector() { std::cout << "~TestDetector" << std::endl; };
 };
