+++ src/modules/bifrost/geometry/BifrostGeometry.cpp
@@ -9,7 +9,7 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include <caen/geometry/BifrostGeometry.h>
+#include <modules/bifrost/geometry/BifrostGeometry.h>
 
 // #undef TRC_LEVEL
 // #define TRC_LEVEL TRC_L_DEB
