+++ src/modules/caen/CaenInstrument.h
@@ -13,12 +13,12 @@
 
 #include <caen/CaenBase.h> // to get CaenSettings
 #include <caen/Counters.h>
-#include <caen/geometry/BifrostGeometry.h>
+#include <bifrost/geometry/BifrostGeometry.h>
 #include <caen/geometry/Calibration.h>
 #include <caen/geometry/Config.h>
-#include <caen/geometry/LokiGeometry.h>
-#include <caen/geometry/MiraclesGeometry.h>
-#include <caen/geometry/PanelGeometry.h>
+#include <loki/geometry/LokiGeometry.h>
+#include <miracles/geometry/MiraclesGeometry.h>
+#include <loki/geometry/PanelGeometry.h>
 #include <caen/readout/Readout.h>
 #include <common/readout/ess/ESSTime.h>
 #include <common/readout/ess/Parser.h>
