+++ src/modules/caen/geometry/Config.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include <caen/geometry/PanelGeometry.h>
+#include <loki/geometry/PanelGeometry.h>
 #include <common/debug/Trace.h>
 #include <logical_geometry/ESSGeometry.h>
 #include <string>
