+++ src/modules/caen/test/CMakeLists.txt
@@ -11,71 +11,6 @@ get_filename_component(CAEN_FULL "${ESS_MODULE_DIR}/caen/configs/Loki.json" ABSO
 ## Pure unit tests
 ##
 
-##
-## Analog geometry - amplitude calculations
-##
-set(LokiGeometryTest_INC 
-    ../geometry/LokiGeometry.h
-    ../geometry/PanelGeometry.h
-    ../geometry/Config.h
-    ../geometry/Calibration.h
-)
-set(LokiGeometryTest_SRC 
-    LokiGeometryTest.cpp
-    ../geometry/LokiGeometry.cpp
-    ../geometry/Config.cpp
-    ../geometry/Calibration.cpp
-)
-create_test_executable(LokiGeometryTest)
-
-set(BifrostGeometryTest_INC 
-    ../geometry/BifrostGeometry.h
-    ../geometry/Config.h
-    ../geometry/Calibration.h
-)
-set(BifrostGeometryTest_SRC 
-    BifrostGeometryTest.cpp
-    ../geometry/BifrostGeometry.cpp
-    ../geometry/Config.cpp
-    ../geometry/Calibration.cpp
-)
-create_test_executable(BifrostGeometryTest)
-
-set(MiraclesGeometryTest_INC 
-    ../geometry/MiraclesGeometry.h
-    ../geometry/Config.h
-    ../geometry/Calibration.h
-)
-set(MiraclesGeometryTest_SRC 
-    MiraclesGeometryTest.cpp
-    ../geometry/MiraclesGeometry.cpp
-    ../geometry/Config.cpp
-    ../geometry/Calibration.cpp
-)
-create_test_executable(MiraclesGeometryTest)
-
-##
-## Panel geometry
-##
-set(CaenPanelGeometryTest_INC
-  ../geometry/LokiGeometry.h
-  ../geometry/PanelGeometry.h)
-set(CaenPanelGeometryTest_SRC PanelGeometryTest.cpp)
-create_test_executable(CaenPanelGeometryTest)
-
-##
-## Validate LoKI geometry between EFU and Nexus Constructor
-##
-set(CaenFullGeometryTest_INC
-  LokiFullGeometryTestData.h
-  ../geometry/LokiGeometry.h
-  ../geometry/Calibration.h
-  ../geometry/PanelGeometry.h)
-set(LokiFullGeometryTest_SRC 
-  LokiFullGeometryTest.cpp
-  ../geometry/Calibration.cpp)
-create_test_executable(LokiFullGeometryTest)
-
 
 set(CaenCalibrationTest_INC
   ${ESS_SOURCE_DIR}/common/testutils/SaveBuffer.h
